refactor orm stuff into its own file
